pyproject.toml
[project]
name = "cryptobot"
version = "0.1.0"
description = "加密货币合约量化交易 AI 团队系统 - 信号桥接层"
requires-python = ">=3.12,<3.13"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "pandas>=2.2",
    "numpy>=1.26",
    "TA-Lib>=0.5",
    "pyarrow>=15.0",
    "rich>=13.7",
    "ccxt>=4.0",
    "langgraph>=0.2",
    "apscheduler>=3.10,<4",
]

[project.scripts]
cryptobot = "cryptobot.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cryptobot"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "network: tests requiring network access",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.0.0",
    "ruff>=0.8",
]
freqtrade = [
    "freqtrade>=2024.1",
]
